refactor: postlistの分離
